requirements.txt
google-cloud-speech==2.27.0
srt==3.5.3
tqdm==4.66.5
pytest==8.3.3
